requirements.txt
mss==10.0.0
opencv-python==4.11.0.86
pillow==11.1.0
pyautogui==0.9.54
pytesseract==0.3.13